Refactor: Move visualization logic to a separate module
- Created `visualization.py` to house all visualization-related functions.
- Moved `visualize_policy`, `draw_frozen_lake_map`, `create_q_value_comparison_animation`, and plotting logic for learning progress into `visualization.py`.
- Updated `frozen_lake_q_learning.py` to import and use the functions from the new `visualization` module.
- Ensured all necessary imports (matplotlib, numpy) are handled correctly in the new module.
- Verified the script runs correctly and produces the same output and artifacts as before the refactoring.
