+++ frozen_lake_q_learning.py
@@ -1,9 +1,8 @@
 # Import necessary libraries
 import gymnasium as gym  # OpenAI Gymnasium for the environment
 import numpy as np      # NumPy for numerical operations (Q-table)
-import matplotlib.pyplot as plt # Matplotlib for plotting results
-import matplotlib.animation as animation
-import matplotlib.patches as patches # Moved import to top for convention
+# Import visualization functions from the new module
+import visualization as viz
 
 # --- Environment Initialization ---
 # Same environment for both algorithms
@@ -127,303 +126,35 @@ for algo_name, data in results.items():
       print(f"{count}: {str(sum(r / 1000))}")
       count += 1000
 
-# --- Policy Visualization Function ---
-def visualize_policy(q_table_to_plot, algorithm_name, map_name="4x4"):
-  """
-  Visualizes the learned policy from the Q-table as a grid of actions.
-  Args:
-    q_table_to_plot (np.array): The Q-table containing learned values.
-    algorithm_name (str): Name of the algorithm (e.g., "Q-learning", "SARSA").
-    map_name (str): The name of the map ("4x4" or "8x8") to get grid description.
-  """
-  if map_name == "4x4":
-    desc = ["SFFF", "FHFH", "FFFH", "HFFG"]
-    grid_size = 4
-  elif map_name == "8x8":
-    desc = ["SFFFFFFF", "FFFFFFFF", "FFFHFFFF", "FFFFFHFF",
-            "FFFHFFFF", "FHHFFFHF", "FHFFHFHF", "FFFHFFFG"]
-    grid_size = 8
-  else:
-    print(f"Warning: Map name not supported for detailed visualization ({algorithm_name}). Using generic state numbers.")
-    best_actions_fallback = np.argmax(q_table_to_plot, axis=1)
-    print(f"\nLearned Policy for {algorithm_name} (best action for each state number):")
-    for s, a in enumerate(best_actions_fallback):
-      print(f"State {s}: Action {a}")
-    return
-
-  best_actions = np.argmax(q_table_to_plot, axis=1)
-  policy_grid = best_actions.reshape((grid_size, grid_size))
-  action_symbols = {0: '<', 1: 'v', 2: '>', 3: '^'}
-
-  print(f"\nLearned Policy for {algorithm_name.upper()} (best action for each state on the grid):")
-  print("S: Start, F: Frozen, H: Hole, G: Goal")
-  print("Actions: <: Left, v: Down, >: Right, ^: Up\n")
-  for i in range(grid_size):
-    row_str = ""
-    for j in range(grid_size):
-      state_index = i * grid_size + j
-      state_char = desc[i][j]
-      if state_char in "GH":
-        row_str += state_char + "  "
-      else:
-        action = policy_grid[i, j]
-        row_str += action_symbols.get(action, '?') + "  "
-    print(row_str)
-
 # Visualize policies
-visualize_policy(results['q_learning']['q_table'], "Q-learning", map_name="4x4")
-visualize_policy(results['sarsa']['q_table'], "SARSA", map_name="4x4")
+viz.visualize_policy(results['q_learning']['q_table'], "Q-learning", map_name="4x4")
+viz.visualize_policy(results['sarsa']['q_table'], "SARSA", map_name="4x4")
 
 
 # --- Function to Draw FrozenLake Map ---
-def draw_frozen_lake_map(ax, map_desc):
-    """
-    Draws the FrozenLake map on a given matplotlib Axes object.
-
-    Args:
-        ax (matplotlib.axes.Axes): The axes object to draw on.
-        map_desc (list of str): Description of the map (e.g., ["SFFF", "FHFH", ...]).
-    """
-    grid_size = len(map_desc)
-    ax.set_xlim(-0.5, grid_size - 0.5)
-    ax.set_ylim(grid_size - 0.5, -0.5) # Inverted y-axis for matrix representation
-    ax.set_xticks(np.arange(grid_size))
-    ax.set_yticks(np.arange(grid_size))
-    ax.set_xticklabels([]) # Hide x-axis numbers
-    ax.set_yticklabels([]) # Hide y-axis numbers
-    # ax.grid(True, which='both', color='black', linewidth=1) # Removed grid lines
-    ax.set_aspect('equal', adjustable='box')
-    ax.set_title("FrozenLake Map")
-
-    colors = {
-        'S': 'lightblue',
-        'F': 'lightgrey',
-        'H': 'black',
-        'G': 'lightgreen'
-    }
-    text_colors = {
-        'S': 'black',
-        'F': 'black',
-        'H': 'white',
-        'G': 'black'
-    }
-
-    for r, row_str in enumerate(map_desc):
-        for c, char in enumerate(row_str):
-            rect = patches.Rectangle((c - 0.5, r - 0.5), 1, 1, linewidth=1, edgecolor='black',
-                                     facecolor=colors.get(char, 'white'))
-            ax.add_patch(rect)
-            ax.text(c, r, char, ha="center", va="center", color=text_colors.get(char, 'black'), fontsize=10)
+# ... (visualization code removed, handled by viz.draw_frozen_lake_map) ...
 
 # --- Q-value Heatmap Animation Function ---
-def create_q_value_comparison_animation(
-    q_tables_history_algo1,
-    q_tables_history_algo2,
-    num_s,
-    num_a,
-    algo1_name,
-    algo2_name,
-    filename="q_values_evolution_comparison.gif"
-):
-    """
-    Creates and saves a GIF animation comparing the Q-table evolution of two algorithms side-by-side.
-    """
-    if not q_tables_history_algo1:
-        print(f"Q-table history for {algo1_name} is empty. Cannot create animation.")
-        return
-    if not q_tables_history_algo2:
-        print(f"Q-table history for {algo2_name} is empty. Cannot create animation.")
-        return
-    if len(q_tables_history_algo1) != len(q_tables_history_algo2):
-        print("Warning: Q-table history lengths differ. Animation will be truncated to the shorter history.")
-        min_len = min(len(q_tables_history_algo1), len(q_tables_history_algo2))
-        q_tables_history_algo1 = q_tables_history_algo1[:min_len]
-        q_tables_history_algo2 = q_tables_history_algo2[:min_len]
-
-    # import matplotlib.patches as patches # This line moved to top of file
-
-    fig, axes = plt.subplots(1, 3, figsize=(20, 8)) # Changed from 1,2 to 1,3 and adjusted figsize
-
-    # Define the map description for FrozenLake 4x4
-    map_desc = ["SFFF", "FHFH", "FFFH", "HFFG"]
-    # Draw the FrozenLake map on the first subplot
-    draw_frozen_lake_map(axes[0], map_desc)
-
-    # Determine global Q-value range for consistent color scaling for heatmaps
-    # Use only heatmap axes (axes[1] and axes[2]) for Q-value related setup
-    all_q_tables = q_tables_history_algo1 + q_tables_history_algo2
-    global_min_q = np.min([np.min(q_table) for q_table in all_q_tables if q_table.size > 0])
-    global_max_q = np.max([np.max(q_table) for q_table in all_q_tables if q_table.size > 0])
-
-    if global_min_q == global_max_q: # Avoid issues with norm if all values are same
-        global_min_q -= 0.1
-        global_max_q += 0.1
-
-    norm = plt.Normalize(vmin=global_min_q, vmax=global_max_q)
-    cmap = plt.cm.viridis
-
-    grid_size = int(np.sqrt(num_s)) # Assuming a square grid, e.g., 4 for 16 states
-
-    # Action to vertex calculation helper
-    def get_triangle_vertices(col, row, action): # col, row are cell indices (center of cell in plot coords)
-        # Corrected triangle orientations
-        # Action mapping: 0: Left, 1: Down, 2: Right, 3: Up
-        # y-axis is inverted: smaller y is UP, larger y is DOWN
-        # x-axis is standard: smaller x is LEFT, larger x is RIGHT
-        # Cell center is (col, row) in plot coordinates. Cell spans [col-0.5, col+0.5] and [row-0.5, row+0.5].
-
-        half = 0.5
-        if action == 3: # Up (North) - Triangle should be at the TOP of the cell, pointing UP (towards smaller y)
-            # Base is top edge of cell, apex is center.
-            # Top edge y-coordinate: row - 0.5
-            # Vertices: [(col - 0.5, row - 0.5), (col + 0.5, row - 0.5), (col, row)]
-            return [(col - half, row - half), (col + half, row - half), (col, row)]
-        elif action == 1: # Down (South) - Triangle should be at the BOTTOM of the cell, pointing DOWN (towards larger y)
-            # Base is bottom edge of cell, apex is center.
-            # Bottom edge y-coordinate: row + 0.5
-            # Vertices: [(col - 0.5, row + 0.5), (col + 0.5, row + 0.5), (col, row)]
-            return [(col - half, row + half), (col + half, row + half), (col, row)]
-        elif action == 0: # Left (West) - Triangle should be at the LEFT of the cell, pointing LEFT (towards smaller x)
-            # Base is left edge of cell, apex is center.
-            # Left edge x-coordinate: col - 0.5
-            # Vertices: [(col - 0.5, row - 0.5), (col - 0.5, row + 0.5), (col, row)]
-            return [(col - half, row - half), (col - half, row + half), (col, row)]
-        elif action == 2: # Right (East) - Triangle should be at the RIGHT of the cell, pointing RIGHT (towards larger x)
-            # Base is right edge of cell, apex is center.
-            # Right edge x-coordinate: col + 0.5
-            # Vertices: [(col + 0.5, row - 0.5), (col + 0.5, row + 0.5), (col, row)]
-            return [(col + half, row - half), (col + half, row + half), (col, row)]
-        return []
-
-
-    # Setup for heatmap axes (axes[1] and axes[2])
-    for i, ax_heatmap in enumerate(axes[1:]): # Iterate only over heatmap axes
-        ax_heatmap.set_xlim(-0.5, grid_size - 0.5)
-        ax_heatmap.set_ylim(grid_size - 0.5, -0.5) # Inverted y-axis for matrix-like display
-        ax_heatmap.set_xticks(np.arange(grid_size))
-        ax_heatmap.set_yticks(np.arange(grid_size))
-        ax_heatmap.set_xlabel("Column")
-        ax_heatmap.set_ylabel("Row")
-        ax_heatmap.set_aspect('equal', adjustable='box')
-        ax_heatmap.grid(True, which='both', color='grey', linewidth=0.5, linestyle='--')
-
-    # Add a single colorbar for the heatmaps
-    sm = plt.cm.ScalarMappable(cmap=cmap, norm=norm)
-    sm.set_array([])
-    # Position colorbar relative to the heatmap axes
-    fig.colorbar(sm, ax=axes[1:].ravel().tolist(), label="Q-value", aspect=30, pad=0.02, location='right')
-
-
-    def update(frame_number):
-        q_table_snapshot1 = q_tables_history_algo1[frame_number]
-        q_table_snapshot2 = q_tables_history_algo2[frame_number]
-
-        # Clear only heatmap axes
-        axes[1].clear()
-        axes[2].clear()
-
-        # Re-apply settings after clearing for heatmap axes
-        for i_ax_clear, ax_clear in enumerate(axes[1:]): # Iterate only over heatmap axes
-            ax_clear.set_xlim(-0.5, grid_size - 0.5)
-            ax_clear.set_ylim(grid_size - 0.5, -0.5)
-            ax_clear.set_xticks(np.arange(grid_size))
-            ax_clear.set_yticks(np.arange(grid_size))
-            ax_clear.set_xlabel("Column")
-            ax_clear.set_ylabel("Row")
-            ax_clear.set_aspect('equal', adjustable='box')
-            ax_clear.grid(True, which='both', color='grey', linewidth=0.5, linestyle='--')
-            if i_ax_clear == 0: # This is now axes[1]
-                 ax_clear.set_title(f"{algo1_name} Q-values at Episode {(frame_number + 1) * q_table_log_interval}")
-            else: # This is now axes[2]
-                 ax_clear.set_title(f"{algo2_name} Q-values at Episode {(frame_number + 1) * q_table_log_interval}")
-
-        current_artists_ax1 = [] # For axes[1]
-        current_artists_ax2 = [] # For axes[2]
-
-        # Process Algorithm 1 (on axes[1])
-        for state in range(num_s):
-            row = state // grid_size
-            col = state % grid_size
-            for action in range(num_a):
-                q_value = q_table_snapshot1[state, action]
-                # Note: Original get_triangle_vertices might need y-coordinates flipped if using matrix indexing (row 0 at top)
-                # The current get_triangle_vertices is adjusted for inverted y-axis (ylim(bottom, top))
-                vertices = get_triangle_vertices(col, row, action) # row, col from matrix perspective
-                color = cmap(norm(q_value))
-                polygon = patches.Polygon(vertices, closed=True, facecolor=color, edgecolor='black', linewidth=0.5)
-                axes[1].add_patch(polygon)
-                current_artists_ax1.append(polygon)
-
-        # Process Algorithm 2 (on axes[2])
-        for state in range(num_s):
-            row = state // grid_size
-            col = state % grid_size
-            for action in range(num_a):
-                q_value = q_table_snapshot2[state, action]
-                vertices = get_triangle_vertices(col, row, action) # row, col from matrix perspective
-                color = cmap(norm(q_value))
-                polygon = patches.Polygon(vertices, closed=True, facecolor=color, edgecolor='black', linewidth=0.5)
-                axes[2].add_patch(polygon)
-                current_artists_ax2.append(polygon)
-
-        # The artists lists are mainly for blitting, which is False.
-        # If blit=True, these would need to be returned.
-        return current_artists_ax1 + current_artists_ax2
-
-    ani = animation.FuncAnimation(fig, update, frames=len(q_tables_history_algo1), interval=200, blit=False)
-    try:
-        ani.save(filename, writer='pillow', fps=5)
-        print(f"\nComparison Q-value heatmap animation saved as {filename}")
-    except Exception as e:
-        print(f"Error saving comparison animation: {e}")
-        print("You might need to install a writer like Pillow: pip install Pillow")
-    finally:
-        plt.close(fig)
+# ... (visualization code removed, handled by viz.create_q_value_comparison_animation) ...
 
 # Create and save Q-value animations
 if results['q_learning']['q_table_history'] and results['sarsa']['q_table_history']:
-    create_q_value_comparison_animation(
+    viz.create_q_value_comparison_animation(
         results['q_learning']['q_table_history'],
         results['sarsa']['q_table_history'],
         num_states,
         num_actions,
         "Q-learning",
-        "SARSA"
+        "SARSA",
+        q_table_log_interval # Pass q_table_log_interval
     )
 else:
     print("Skipping comparison animation as Q-table history for one or both algorithms is missing.")
 
 
 # --- Plotting Learning Progress ---
-plt.figure(figsize=(14, 7))
-
-# 1. Plot Raw Rewards per Episode (Q-learning vs SARSA)
-plt.subplot(1, 2, 1) # 1 row, 2 columns, 1st subplot
-plt.plot(results['q_learning']['rewards_all_episodes'], label='Q-learning', alpha=0.7)
-plt.plot(results['sarsa']['rewards_all_episodes'], label='SARSA', alpha=0.7)
-plt.xlabel('Episode')
-plt.ylabel('Reward')
-plt.title('Rewards per Episode (Raw)')
-plt.legend()
-plt.grid(True)
-
-# 2. Plot Moving Average of Rewards (Q-learning vs SARSA)
-plt.subplot(1, 2, 2) # 1 row, 2 columns, 2nd subplot
-moving_avg_window = 100
-q_learning_moving_avg = np.convolve(results['q_learning']['rewards_all_episodes'], np.ones(moving_avg_window)/moving_avg_window, mode='valid')
-sarsa_moving_avg = np.convolve(results['sarsa']['rewards_all_episodes'], np.ones(moving_avg_window)/moving_avg_window, mode='valid')
-
-plt.plot(range(moving_avg_window -1, num_episodes), q_learning_moving_avg, label='Q-learning (Avg)')
-plt.plot(range(moving_avg_window -1, num_episodes), sarsa_moving_avg, label='SARSA (Avg)')
-plt.xlabel(f'Episode (averaged over {moving_avg_window} episodes)')
-plt.ylabel('Average Reward')
-plt.title('Moving Average of Rewards')
-plt.legend()
-plt.grid(True)
-
-plt.tight_layout() # Adjust layout to prevent overlapping titles/labels
-plt.savefig('rewards_comparison.png')
-print("\nTraining finished. Combined rewards plot saved as 'rewards_comparison.png'.\n")
+# ... (visualization code removed, handled by viz.plot_learning_progress) ...
+viz.plot_learning_progress(results, num_episodes)
 
 
 # --- Testing the Learned Policies ---
